+++ packages/react-native/ios/BugsnagReactNative/BugsnagReactNative.mm
@@ -73,6 +73,11 @@ BSG_EXPORT_METHOD(updateContext:(NSString *)context) {
     BSG_EXPORT_RETURN
 }
 
+BSG_EXPORT_METHOD(updateGroupingDiscriminator:(NSString *)groupingDiscriminator) {
+    [Bugsnag setGroupingDiscriminator:groupingDiscriminator];
+    BSG_EXPORT_RETURN
+}
+
 BSG_EXPORT_METHOD(updateCodeBundleId:(NSString *)codeBundleId) {
     Bugsnag.client.codeBundleId = codeBundleId; 
     BSG_EXPORT_RETURN