+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - Add additional grouping discriminator property to events [#2544](https://github.com/bugsnag/bugsnag-js/pull/2544)
+- (react-native) Handle additional grouping discriminator [#2557](https://github.com/bugsnag/bugsnag-js/pull/2557)
 
 ### Changed